OMAC: Fix for calculation of next slot for discovery

+++ MicroFrameworkPK_v4_3/DeviceCode/Include/Samraksh/MAC/OMAC/DiscoveryHandler.h
@@ -81,7 +81,7 @@ class DiscoveryHandler: public EventHandler {
 	void BeaconAckHandler(Message_15_4_t* msg, UINT8 len, NetOpStatus success);
 
 	UINT64 NextEvent(UINT32 currentSlotNum);
-  	UINT16 NextEventinSlots(UINT32 currentSlotNum);
+  	UINT64 NextEventinSlots(UINT32 currentSlotNum);
 	//UINT64 NextEvent(UINT64 currentTicks);
   	void ExecuteEvent(UINT32 currentSlotNum);
   	UINT8 ExecuteEventDone();
